[CSJHW-7]: deleted warnings and cleanup

+++ src/Ozon.Route256.Kafka.OrderEventConsumer.Infrastructure/Migrations/20240417122100_Initial.cs
@@ -21,7 +21,7 @@ create table if not exists sales
       seller_id           bigint not null
     , item_id             bigint not null
     , currency            text not null
-    , sales               bigint not null
+    , sales               decimal not null
     , constraint pk_sales primary key (seller_id, item_id)
 );
 ";
